cdk_infra/requirements.txt
aws-cdk.aws-lambda==1.116.0
aws-cdk.core==1.116.0
aws-cdk.aws-iam==1.116.0
aws-cdk.aws-lambda-go==1.116.0
aws-cdk.aws-s3-notifications==1.116.0
aws-cdk.aws-lambda-event-sources==1.116.0
aws-cdk.aws-events-targets==1.116.0